+++ CHANGES.rst
@@ -1,8 +1,11 @@
 Changelog
 =========
 
-1.3.9 (unreleased)
+1.3.9 (2015-03-26)
 ------------------
+
+- LanguageTool/plone.app.multilingual migration
+  [bloodbare]
 
 - Update tests after removal of ``allowAnonymousViewAbout`` and
   ``validate_email`` properties in CMFPlone.